src/third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-addIceCandidate-timing.https-expected.txt

This is a testharness.js-based test.
PASS addIceCandidate is not resolved first if 2x setLocalDescription operations are pending
FAIL addIceCandidate and setLocalDescription are resolved in the correct order, as defined by the operations chain specification assert_array_equals: expected property 1 to be "setLocalDescription 2" but got "addIceCandidate" (expected array ["setLocalDescription 1", "setLocalDescription 2", "addIceCandidate"] got ["setLocalDescription 1", "addIceCandidate", "setLocalDescription 2"])
PASS onicecandidate fires after resolving setLocalDescription in offerer
PASS onicecandidate fires after resolving setLocalDescription in answerer
Harness: the test ran to completion.
